+++ src/user_input.rs
@@ -28,15 +28,15 @@ pub enum UserInput {
 }
 
 impl UserInput {
-    /// Creates a [`UserInput::Chord`] from an iterator of [`Button`]s
+    /// Creates a [`UserInput::Chord`] from an iterator of [`InputKind`]s
     ///
-    /// If `buttons` has a length of 1, a [`UserInput::Single`] variant will be returned instead.
-    pub fn chord(buttons: impl IntoIterator<Item = impl Into<InputKind>>) -> Self {
+    /// If `inputs` has a length of 1, a [`UserInput::Single`] variant will be returned instead.
+    pub fn chord(inputs: impl IntoIterator<Item = impl Into<InputKind>>) -> Self {
         // We can't just check the length unless we add an ExactSizeIterator bound :(
         let mut length: u8 = 0;
 
         let mut set: PetitSet<InputKind, 8> = PetitSet::default();
-        for button in buttons {
+        for button in inputs {
             length += 1;
             set.insert(button.into());
         }
@@ -122,51 +122,47 @@ impl UserInput {
     }
 
     /// Returns the raw inputs that make up this [`UserInput`]
-    pub fn raw_inputs(
-        &self,
-    ) -> (
-        Vec<GamepadButtonType>,
-        Vec<(AxisType, Option<f32>)>,
-        Vec<KeyCode>,
-        Vec<MouseButton>,
-    ) {
-        let mut axis_data: Vec<(AxisType, Option<f32>)> = Vec::default();
-        let mut gamepad_buttons: Vec<GamepadButtonType> = Vec::default();
-        let mut keyboard_buttons: Vec<KeyCode> = Vec::default();
-        let mut mouse_buttons: Vec<MouseButton> = Vec::default();
-        let mut mouse_wheel: Vec<MouseWheelDirection> = Vec::default();
-        let mut mouse_motion: Vec<MouseMotionDirection> = Vec::default();
+    pub fn raw_inputs(&self) -> RawInputs {
+        let mut raw_inputs = RawInputs::default();
 
         match self {
             UserInput::Single(button) => match *button {
                 InputKind::DualAxis(dual_axis) => {
-                    axis_data.push((dual_axis.x.axis_type, dual_axis.x.value));
-                    axis_data.push((dual_axis.y.axis_type, dual_axis.y.value));
+                    raw_inputs
+                        .axis_data
+                        .push((dual_axis.x.axis_type, dual_axis.x.value));
+                    raw_inputs
+                        .axis_data
+                        .push((dual_axis.y.axis_type, dual_axis.y.value));
                 }
-                InputKind::SingleAxis(single_axis) => {
-                    axis_data.push((single_axis.axis_type, single_axis.value))
-                }
-                InputKind::GamepadButton(button) => gamepad_buttons.push(button),
-                InputKind::Keyboard(button) => keyboard_buttons.push(button),
-                InputKind::Mouse(button) => mouse_buttons.push(button),
-                InputKind::MouseWheel(button) => mouse_wheel.push(button),
-                InputKind::MouseMotion(button) => mouse_motion.push(button),
+                InputKind::SingleAxis(single_axis) => raw_inputs
+                    .axis_data
+                    .push((single_axis.axis_type, single_axis.value)),
+                InputKind::GamepadButton(button) => raw_inputs.gamepad_buttons.push(button),
+                InputKind::Keyboard(button) => raw_inputs.keycodes.push(button),
+                InputKind::Mouse(button) => raw_inputs.mouse_buttons.push(button),
+                InputKind::MouseWheel(button) => raw_inputs.mouse_wheel.push(button),
+                InputKind::MouseMotion(button) => raw_inputs.mouse_motion.push(button),
             },
             UserInput::Chord(button_set) => {
                 for button in button_set.iter() {
                     match *button {
                         InputKind::DualAxis(dual_axis) => {
-                            axis_data.push((dual_axis.x.axis_type, dual_axis.x.value));
-                            axis_data.push((dual_axis.y.axis_type, dual_axis.y.value));
+                            raw_inputs
+                                .axis_data
+                                .push((dual_axis.x.axis_type, dual_axis.x.value));
+                            raw_inputs
+                                .axis_data
+                                .push((dual_axis.y.axis_type, dual_axis.y.value));
                         }
-                        InputKind::SingleAxis(single_axis) => {
-                            axis_data.push((single_axis.axis_type, single_axis.value))
-                        }
-                        InputKind::GamepadButton(button) => gamepad_buttons.push(button),
-                        InputKind::Keyboard(button) => keyboard_buttons.push(button),
-                        InputKind::Mouse(button) => mouse_buttons.push(button),
-                        InputKind::MouseWheel(button) => mouse_wheel.push(button),
-                        InputKind::MouseMotion(button) => mouse_motion.push(button),
+                        InputKind::SingleAxis(single_axis) => raw_inputs
+                            .axis_data
+                            .push((single_axis.axis_type, single_axis.value)),
+                        InputKind::GamepadButton(button) => raw_inputs.gamepad_buttons.push(button),
+                        InputKind::Keyboard(button) => raw_inputs.keycodes.push(button),
+                        InputKind::Mouse(button) => raw_inputs.mouse_buttons.push(button),
+                        InputKind::MouseWheel(button) => raw_inputs.mouse_wheel.push(button),
+                        InputKind::MouseMotion(button) => raw_inputs.mouse_motion.push(button),
                     }
                 }
             }
@@ -179,23 +175,27 @@ impl UserInput {
                 for button in [up, down, left, right] {
                     match *button {
                         InputKind::DualAxis(dual_axis) => {
-                            axis_data.push((dual_axis.x.axis_type, dual_axis.x.value));
-                            axis_data.push((dual_axis.y.axis_type, dual_axis.y.value));
+                            raw_inputs
+                                .axis_data
+                                .push((dual_axis.x.axis_type, dual_axis.x.value));
+                            raw_inputs
+                                .axis_data
+                                .push((dual_axis.y.axis_type, dual_axis.y.value));
                         }
-                        InputKind::SingleAxis(single_axis) => {
-                            axis_data.push((single_axis.axis_type, single_axis.value))
-                        }
-                        InputKind::GamepadButton(button) => gamepad_buttons.push(button),
-                        InputKind::Keyboard(button) => keyboard_buttons.push(button),
-                        InputKind::Mouse(button) => mouse_buttons.push(button),
-                        InputKind::MouseWheel(button) => mouse_wheel.push(button),
-                        InputKind::MouseMotion(button) => mouse_motion.push(button),
+                        InputKind::SingleAxis(single_axis) => raw_inputs
+                            .axis_data
+                            .push((single_axis.axis_type, single_axis.value)),
+                        InputKind::GamepadButton(button) => raw_inputs.gamepad_buttons.push(button),
+                        InputKind::Keyboard(button) => raw_inputs.keycodes.push(button),
+                        InputKind::Mouse(button) => raw_inputs.mouse_buttons.push(button),
+                        InputKind::MouseWheel(button) => raw_inputs.mouse_wheel.push(button),
+                        InputKind::MouseMotion(button) => raw_inputs.mouse_motion.push(button),
                     }
                 }
             }
         };
 
-        (gamepad_buttons, axis_data, keyboard_buttons, mouse_buttons)
+        raw_inputs
     }
 }
 
@@ -319,5 +319,255 @@ impl From<MouseWheelDirection> for InputKind {
 impl From<MouseMotionDirection> for InputKind {
     fn from(input: MouseMotionDirection) -> Self {
         InputKind::MouseMotion(input)
+    }
+}
+
+/// The basic input events that make up a [`UserInput`].
+///
+/// Obtained by calling [`UserInput::raw_inputs()`].
+#[derive(Default, Debug, Clone, PartialEq)]
+pub struct RawInputs {
+    /// Physical keyboard buttons
+    pub keycodes: Vec<KeyCode>,
+    /// Mouse buttons
+    pub mouse_buttons: Vec<MouseButton>,
+    /// Discretized mouse wheel inputs
+    pub mouse_wheel: Vec<MouseWheelDirection>,
+    /// Discretized mouse motion inputs
+    pub mouse_motion: Vec<MouseMotionDirection>,
+    /// Gamepad buttons, independent of a [`Gamepad`](bevy::input::gamepad::Gamepad)
+    pub gamepad_buttons: Vec<GamepadButtonType>,
+    /// Axis-like data
+    ///
+    /// The `f32` stores the magnitude of the axis motion, and is only used for input mocking.
+    pub axis_data: Vec<(AxisType, Option<f32>)>,
+}
+
+#[cfg(test)]
+impl RawInputs {
+    fn from_keycode(keycode: KeyCode) -> RawInputs {
+        RawInputs {
+            keycodes: vec![keycode],
+            ..Default::default()
+        }
+    }
+
+    fn from_mouse_button(mouse_button: MouseButton) -> RawInputs {
+        RawInputs {
+            mouse_buttons: vec![mouse_button],
+            ..Default::default()
+        }
+    }
+
+    fn from_gamepad_button(gamepad_button: GamepadButtonType) -> RawInputs {
+        RawInputs {
+            gamepad_buttons: vec![gamepad_button],
+            ..Default::default()
+        }
+    }
+
+    fn from_mouse_wheel(direction: MouseWheelDirection) -> RawInputs {
+        RawInputs {
+            mouse_wheel: vec![direction],
+            ..Default::default()
+        }
+    }
+
+    fn from_mouse_direction(direction: MouseMotionDirection) -> RawInputs {
+        RawInputs {
+            mouse_motion: vec![direction],
+            ..Default::default()
+        }
+    }
+
+    fn from_dual_axis(axis: DualAxis) -> RawInputs {
+        RawInputs {
+            axis_data: vec![
+                (axis.x.axis_type, axis.x.value),
+                (axis.y.axis_type, axis.y.value),
+            ],
+            ..Default::default()
+        }
+    }
+
+    fn from_single_axis(axis: SingleAxis) -> RawInputs {
+        RawInputs {
+            axis_data: vec![(axis.axis_type, axis.value)],
+            ..Default::default()
+        }
+    }
+}
+
+#[cfg(test)]
+mod raw_input_tests {
+    use crate::{
+        axislike::AxisType,
+        user_input::{InputKind, RawInputs, UserInput},
+    };
+
+    #[test]
+    fn simple_chord() {
+        use bevy::input::gamepad::GamepadButtonType;
+
+        let buttons = vec![GamepadButtonType::Start, GamepadButtonType::Select];
+        let raw_inputs = UserInput::chord(buttons.clone()).raw_inputs();
+        let expected = RawInputs {
+            gamepad_buttons: buttons,
+            ..Default::default()
+        };
+
+        assert_eq!(expected, raw_inputs);
+    }
+
+    #[test]
+    fn mixed_chord() {
+        use crate::axislike::SingleAxis;
+        use bevy::input::gamepad::GamepadAxisType;
+        use bevy::input::gamepad::GamepadButtonType;
+
+        let chord = UserInput::chord([
+            InputKind::GamepadButton(GamepadButtonType::Start),
+            InputKind::SingleAxis(SingleAxis::symmetric(GamepadAxisType::LeftZ, 0.)),
+        ]);
+
+        let raw = chord.raw_inputs();
+        let expected = RawInputs {
+            gamepad_buttons: vec![GamepadButtonType::Start],
+            axis_data: vec![(AxisType::Gamepad(GamepadAxisType::LeftZ), None)],
+            ..Default::default()
+        };
+
+        assert_eq!(expected, raw);
+    }
+
+    mod gamepad {
+        use crate::user_input::{RawInputs, UserInput};
+
+        #[test]
+        fn gamepad_button() {
+            use bevy::input::gamepad::GamepadButtonType;
+
+            let button = GamepadButtonType::Start;
+            let expected = RawInputs::from_gamepad_button(button);
+            let raw = UserInput::from(button).raw_inputs();
+            assert_eq!(expected, raw);
+        }
+
+        #[test]
+        fn single_gamepad_axis() {
+            use crate::axislike::SingleAxis;
+            use bevy::input::gamepad::GamepadAxisType;
+
+            let direction = SingleAxis::from_value(GamepadAxisType::LeftStickX, 1.0);
+            let expected = RawInputs::from_single_axis(direction);
+            let raw = UserInput::from(direction).raw_inputs();
+            assert_eq!(expected, raw)
+        }
+
+        #[test]
+        fn dual_gamepad_axis() {
+            use crate::axislike::DualAxis;
+            use bevy::input::gamepad::GamepadAxisType;
+
+            let direction = DualAxis::from_value(
+                GamepadAxisType::LeftStickX,
+                GamepadAxisType::LeftStickY,
+                0.5,
+                0.7,
+            );
+            let expected = RawInputs::from_dual_axis(direction);
+            let raw = UserInput::from(direction).raw_inputs();
+            assert_eq!(expected, raw)
+        }
+    }
+
+    mod keyboard {
+        use crate::user_input::{RawInputs, UserInput};
+
+        #[test]
+        fn keyboard_button() {
+            use bevy::input::keyboard::KeyCode;
+
+            let button = KeyCode::A;
+            let expected = RawInputs::from_keycode(button);
+            let raw = UserInput::from(button).raw_inputs();
+            assert_eq!(expected, raw);
+        }
+    }
+
+    mod mouse {
+        use crate::user_input::{RawInputs, UserInput};
+
+        #[test]
+        fn mouse_button() {
+            use bevy::input::mouse::MouseButton;
+
+            let button = MouseButton::Left;
+            let expected = RawInputs::from_mouse_button(button);
+            let raw = UserInput::from(button).raw_inputs();
+            assert_eq!(expected, raw);
+        }
+
+        #[test]
+        fn mouse_wheel() {
+            use crate::buttonlike::MouseWheelDirection;
+
+            let direction = MouseWheelDirection::Down;
+            let expected = RawInputs::from_mouse_wheel(direction);
+            let raw = UserInput::from(direction).raw_inputs();
+            assert_eq!(expected, raw);
+        }
+
+        #[test]
+        fn mouse_motion() {
+            use crate::buttonlike::MouseMotionDirection;
+
+            let direction = MouseMotionDirection::Up;
+            let expected = RawInputs::from_mouse_direction(direction);
+            let raw = UserInput::from(direction).raw_inputs();
+            assert_eq!(expected, raw);
+        }
+
+        #[test]
+        fn single_mousewheel_axis() {
+            use crate::axislike::{MouseWheelAxisType, SingleAxis};
+
+            let direction = SingleAxis::from_value(MouseWheelAxisType::X, 1.0);
+            let expected = RawInputs::from_single_axis(direction);
+            let raw = UserInput::from(direction).raw_inputs();
+            assert_eq!(expected, raw)
+        }
+
+        #[test]
+        fn dual_mousewheel_axis() {
+            use crate::axislike::{DualAxis, MouseWheelAxisType};
+
+            let direction =
+                DualAxis::from_value(MouseWheelAxisType::X, MouseWheelAxisType::Y, 1.0, 1.0);
+            let expected = RawInputs::from_dual_axis(direction);
+            let raw = UserInput::from(direction).raw_inputs();
+            assert_eq!(expected, raw)
+        }
+
+        #[test]
+        fn single_mouse_motion_axis() {
+            use crate::axislike::{MouseMotionAxisType, SingleAxis};
+
+            let direction = SingleAxis::from_value(MouseMotionAxisType::X, 1.0);
+            let expected = RawInputs::from_single_axis(direction);
+            let raw = UserInput::from(direction).raw_inputs();
+            assert_eq!(expected, raw)
+        }
+
+        #[test]
+        fn dual_mouse_motion_axis() {
+            use crate::axislike::{DualAxis, MouseMotionAxisType};
+
+            let direction =
+                DualAxis::from_value(MouseMotionAxisType::X, MouseMotionAxisType::Y, 1.0, 1.0);
+            let expected = RawInputs::from_dual_axis(direction);
+            let raw = UserInput::from(direction).raw_inputs();
+            assert_eq!(expected, raw)
+        }
     }
 }